correction in classes

+++ src/main/java/br/com/appcinemateca/api/dto/DiretorDTO.java
@@ -23,7 +23,7 @@ public class DiretorDTO implements Serializable {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
